+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Changelog
 * :issue:`22`: Update the :meth:`clustermap` method to accept :class:`pandas.DataFrame` in case the input data was not generated from QIIME 2 (e.g. shotgun sequencing). You can now also flip the x and y axes with the ``flip`` option.
 * :issue:`22`: Add a new main plotting method :meth:`heatmap`.
 * :issue:`24`: Update the :meth:`pname` method to allow returning of more than one tax level.
+* Deprecate the ``add_datapoints`` argument in the :meth:`taxa_abundance_box_plot` method.
 
 1.10.0 (2021-07-06)
 -------------------